fish_pipeline_v6/requirements.txt
torch
opencv-python
albumentations
scikit-learn
ultralytics
Pillow
matplotlib
numpy
